feat(sysload_analyz.sh): add script to periodically check system load and generate perf report if load exceeds threshold
The script `sysload_analyz.sh` is added to periodically check the system load of the system. If the condition `SYSLOAD_1m/CPU_CORES x 100 > 150` is met, it will run and collect a point in time perf trace and generate a flamegraph at 99Hz polling.

The script calculates the system load by dividing the 1-minute load average (`SYSLOAD_1m`) by the number of CPU cores (`CPU_CORES`) and multiplying it by 100. If the calculated percentage load exceeds 120, it will generate a perf report. Otherwise, it will output "sysload is normal".

This script is intended to be run periodically to monitor the system load and take action if necessary.